NP-30787-import-of-same-institution, code review
